reorganize signal module and leave stubs in windows
