+++ lib/clif-backend/src/signal/windows.rs
@@ -1,0 +1,6 @@
+use crate::signal::HandlerData;
+use wasmer_runtime_core::error::RuntimeResult;
+
+pub fn call_protected<T>(handler_data: &HandlerData, f: impl FnOnce() -> T) -> RuntimeResult<T> {
+    unimplemented!("TODO");
+}
